refactor(agentic-org): keep durable state adapter replaceable
