+++ agentic-organization/packages/README.md
@@ -11,8 +11,8 @@ host, or Kubernetes deployment is introduced.
 | ----------------- | -------------------------------------------------------------------------------------------------------------------------- |
 | `domain`          | typed command names, event names, aggregate names, work item state machine, event envelope, shared records                 |
 | `application`     | command pipeline, handler registry, idempotency handling, state-store ports, first supervisor-chain signal command handler |
-| `state`           | in-memory Organization state-store factory used as the first repository port fake                                          |
-| `state-cockroach` | CockroachDB state-store/outbox-source contracts, SQL statement catalogs, and first core-state migration skeleton           |
+| `state`           | generic state-store and outbox-source ports plus the in-memory Organization state-store factory fake                       |
+| `state-cockroach` | first replaceable durable SQL adapter for the state-store/outbox-source ports, backed by CockroachDB                       |
 | `messaging`       | NATS subject contract, outbox publisher port, event publisher port, and domain resolver without a live NATS dependency     |
 | `messaging-nats`  | NATS JetStream event publisher adapter contract with canonical JSON payloads, headers, and message IDs                     |
 | `observability`   | LGTM/OpenTelemetry attribute projection from Agentic event envelopes                                                       |
@@ -38,7 +38,10 @@ supervisor-chain signal command
 CockroachDB, JetStream publishing, Temporal, Dapr, Hermes, Hindsight,
 and the hat-system CRDs come next as adapters behind these contracts.
 They should not redefine command names, event names, state names,
-correlation fields, or policy authority.
+correlation fields, or policy authority. CockroachDB is the first
+durable state adapter because it exists in `full-ai-cluster`; application
+and messaging code must remain database-agnostic so a later durable
+store can replace it behind the same ports.
 
 The application package must not construct concrete state adapters.
 Runtime hosts and tests provide a `CommandStateStoreFactory`; the state
@@ -46,16 +49,17 @@ package implements the current in-memory factory. Command routing uses a
 handler registry so new commands add handlers instead of editing a
 central `switch` or `if` dispatcher.
 
-`CommandStateStore` is async even when backed by the in-memory fake. The
-real CockroachDB adapter must not be squeezed into a synchronous toy
-shape.
+`CommandStateStore` and `OutboxEventSource` are async even when backed
+by in-memory fakes. Durable adapters must not be squeezed into a
+synchronous toy shape.
 
 The outbox publisher owns the generic publish loop: claim unpublished
-outbox rows, resolve a typed Organization domain, publish through an
-`EventPublisher` port, and mark the outbox row published only after the
-publish returns successfully. The NATS package implements that port and
-is the only package in this slice that knows about NATS headers,
-message IDs, and JSON transport payloads.
+outbox rows from the generic state port, resolve a typed Organization
+domain, publish through an `EventPublisher` port, and mark the outbox
+row published only after the publish returns successfully. The NATS
+package implements that publisher port and is the only package in this
+slice that knows about NATS headers, message IDs, and JSON transport
+payloads. State adapters must not import messaging adapters.
 
 ## Validation
