+++ agentic-organization/packages/governance/src/package-dependency-boundaries.ts
@@ -5,6 +5,7 @@ import { fileURLToPath } from "node:url";
 export const PackageBoundaryRule = {
   Application: "application",
   Messaging: "messaging",
+  StateAdapter: "state_adapter",
 } as const;
 
 export type PackageBoundaryRule = (typeof PackageBoundaryRule)[keyof typeof PackageBoundaryRule];
