+++ agentic-organization/packages/messaging/src/index.ts
@@ -4,12 +4,9 @@ export {
   OutboxPublishOutcomeStatus,
   createOutboxPublisher,
   resolveAgenticMessagingDomain,
-  type ClaimUnpublishedOutboxEventsInput,
   type CreateOutboxPublisherInput,
   type EventPublication,
   type EventPublisher,
-  type MarkOutboxEventPublishedInput,
-  type OutboxEventSource,
   type OutboxPublishBatchResult,
   type OutboxPublisher,
   type ResolveAgenticMessagingDomain,
