+++ agentic-organization/packages/state/src/index.ts
@@ -3,3 +3,8 @@ export {
   type InMemoryOrganizationStoreFactory,
   type InMemoryOrganizationStoreSnapshot,
 } from "./in-memory-organization-store.ts";
+export type {
+  ClaimUnpublishedOutboxEventsInput,
+  MarkOutboxEventPublishedInput,
+  OutboxEventSource,
+} from "./outbox-event-source.ts";
